fix(filenames): rename media mention files for consistency
