+++ src/interfaces/graphql/resolvers/index.ts
@@ -2,8 +2,8 @@ import { authMutation as AdminMutation } from './admin/AuthMutation';
 import { blobMutations as BlobMutation } from './blobStorage/blobMutation';
 import { EventsMutation } from './events/EventsMutation';
 import { EventsQuery } from './events/EventsQuery';
-import { MediaMentionsMutation } from './media-mentions/MedeaMentionMutation';
-import { MediaMentionsQuery } from './media-mentions/MedeaMentionQuery';
+import { MediaMentionsMutation } from './media-mentions/MediaMentionMutation';
+import { MediaMentionsQuery } from './media-mentions/MediaMentionQuery';
 import { NewsMutation } from './news/NewsMutation';
 import { NewsQuery } from './news/NewsQuery';
 import { PageMutation } from './page/PageMutation';
