+++ src/interfaces/graphql/resolvers/media-mentions/MediaMentionMutation.test.ts
@@ -1,6 +1,6 @@
 import { GraphQLError } from 'graphql';
 
-import { CreateMediaMentionGQLInput, MediaMentionsMutation, UpdateMediaMentionGQLInput } from './MedeaMentionMutation';
+import { CreateMediaMentionGQLInput, MediaMentionsMutation, UpdateMediaMentionGQLInput } from './MediaMentionMutation';
 import type { GraphQLContext } from '~/back-shared/types/container/types';
 import { MediaMentionEntity } from '~/domain/entities/MediaMentions';
 import type { IMediaMentionsRepository } from '~/domain/repositories/mediaMentionsRepository';
