+++ src/interfaces/graphql/resolvers/media-mentions/MediaMentionQuery.test.ts
@@ -1,4 +1,4 @@
-import { MediaMentionsQuery } from './MedeaMentionQuery';
+import { MediaMentionsQuery } from './MediaMentionQuery';
 import type { GraphQLContext } from '~/back-shared/types/container/types';
 import type { IMediaMentionsRepository } from '~/domain/repositories/mediaMentionsRepository';
 import {  SortOrder } from '~/types/enums/common.enums';
